Extract FacetSearchBuilder
