+++ app/controllers/concerns/blacklight/catalog.rb
@@ -9,6 +9,7 @@ module Blacklight::Catalog
   include Blacklight::Configurable
   include Blacklight::SearchContext
   include Blacklight::Searchable
+  include Blacklight::Facetable
 
   # The following code is executed when someone includes blacklight::catalog in their
   # own controller.
@@ -84,11 +85,11 @@ module Blacklight::Catalog
     raise ActionController::RoutingError, 'Not Found' unless @facet
 
     @response = if params[:query_fragment].present?
-                  search_service.facet_suggest_response(@facet.key, params[:query_fragment])
+                  facet_search_service.facet_suggest_response(@facet.key, params[:query_fragment])
                 else
-                  search_service.facet_field_response(@facet.key)
+                  facet_search_service.facet_field_response(@facet.key)
                 end
-    # @display_facet is a Blacklight::Solr::Response::Facets::FacetField
+
     @display_facet = @response.aggregations[@facet.field]
 
     # @presenter is a Blacklight::FacetFieldPresenter
