+++ app/models/facet_search_builder.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class FacetSearchBuilder < Blacklight::FacetSearchBuilder
+  include Blacklight::Solr::FacetSearchBuilderBehavior
+end
